uapi: convert net/if.h
